+++ comics/comics/lookingforgroup.py
@@ -19,11 +19,13 @@ class Crawler(CrawlerBase):
         feed = self.parse_feed('http://feeds.feedburner.com/LookingForGroup')
         images = []
         for entry in feed.for_date(pub_date):
-            if entry.title.isdigit():
-                url = entry.summary.src('a[rel="bookmark"] img')
-                if url:
-                    url = url.replace('-210x300', '')
-                title = entry.title
-                images.append(CrawlerImage(url, title))
+            if not entry.title.isdigit():
+                continue
+            url = entry.summary.src(
+                'a[rel="bookmark"] img[src*="lfgcomic.com/wp-content/"]')
+            if url:
+                url = url.replace('-210x300', '')
+            title = entry.title
+            images.append(CrawlerImage(url, title))
         if images:
             return images