+++ comics/comics/rocky.py
@@ -12,7 +12,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 14
-    schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
+    schedule = 'Mo,Tu,We,Th,Fr,Sa'
     time_zone = 'Europe/Oslo'
 
     def crawl(self, pub_date):