+++ comics/comics/threewordphrase.py
@@ -7,19 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.threewordphrase.com/'
     start_date = '2010-07-13'
+    end_date = '2014-05-27'
     rights = 'Ryan Pequin'
+    active = False
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 0
-    time_zone = 'US/Pacific'
-
     def crawl(self, pub_date):
-         # Thee feed has broken dates, so we fetch only the latest one
-        feed = self.parse_feed('http://www.threewordphrase.com/rss.xml')
-        if feed.all():
-            entry = feed.all()[0]
-            url = entry.link.replace('.htm', '.gif')
-            title = entry.title
-            text = entry.summary.root.text
-            return CrawlerImage(url, title, text)
+        pass  # Comic no longer published