+++ comics/comics/vgcats.py
@@ -20,7 +20,7 @@ class Crawler(CrawlerBase):
     headers = {'User-Agent': 'Mozilla/4.0'}
 
     def crawl(self, pub_date):
-         # FIXME: Seems like they are using gif images now and then
+        # FIXME: Seems like they are using gif images now and then
         if pub_date < datetime.date(2003, 5, 1):
             file_ext = 'gif'
         else: